Add dummy torrent client
Add the option to use a dummy torrent client that prints to
a file. Useful for testing.

Small bugfix as well. The ArgumentException thrown on invalid
torrent client used a string instead of a symbol as the key
for the options hash when interpolating the client name into
the error message.