Change Parser to enable testing of help message
Parser now always creates an OptionParser and keeps it in an
instance variable. Instead of printing the usage message directly
on -h, --help and invalid switches, options[:action] is set to
:help, and the Client handles printing of the message. The message
can be accessed through the new #help() method on the Parser.

The Client has a corresponding #help() method, making the message
accessible to tests etc.